Refactor: Standardize package name for extensions and update Kover exclusion rules
- Renamed the `extensions` package in `common-feature` to `extension` for consistency.
- Updated import paths in various files to reflect the package name change:
    - `DocumentDetailsTransformer.kt`
    - `RequestTransformer.kt`
    - `TransactionDetailsInteractor.kt`
    - `PresentationSuccessInteractor.kt`
    - `ProximitySuccessInteractor.kt`
    - `DocumentIssuanceSuccessInteractor.kt`
- Modified Kover exclusion rules in `KoverExclusionRules.kt`:
    - Added new common exclusion rules for packages: `DOMAIN_MODELS`, `EXTENSIONS`, `CONFIGS`, `TRANSFORMERS`, `UTILS`, `COMPONENTS`, `PROVIDERS`.
    - Removed specific class and package exclusions from `BusinessLogic`, `UiLogic`, `CommonFeature`, `DashboardFeature`, `ProximityFeature` as they are now covered by the common rules.
    - Updated `UiLogic` to exclude `eu.europa.ec.uilogic.navigation` and `eu.europa.ec.uilogic.serializer` packages.
- Adjusted Kover configuration in `KoverAndroid.kt` by removing explicit `extensions` receiver type.
- Updated comments in `assembly-logic/build.gradle.kts` for clarity regarding Kover reports.
