+++ build-logic/convention/src/main/kotlin/project/convention/logic/kover/KoverExclusionRules.kt
@@ -22,6 +22,13 @@ private const val KOIN = "*.ksp.*"
 private const val BUILD_CONFIG = "*BuildConfig*"
 private const val SCREEN_COMPOSABLES = "*Screen*"
 private const val MODELS = "eu.europa.ec.*.model"
+private const val DOMAIN_MODELS = "eu.europa.ec.*.domain"
+private const val EXTENSIONS = "eu.europa.ec.*.extension"
+private const val CONFIGS = "eu.europa.ec.*.config"
+private const val TRANSFORMERS = "eu.europa.ec.*.transformer"
+private const val COMPONENTS = "eu.europa.ec.*.component"
+private const val UTILS = "eu.europa.ec.*.util"
+private const val PROVIDERS = "eu.europa.ec.*.provider"
 private const val DI = "eu.europa.ec.*.di"
 private const val ROUTER_GRAPH = "eu.europa.ec.*.router"
 
@@ -40,7 +47,7 @@ sealed interface KoverExclusionRules {
     val commonClasses: List<String>
         get() = listOf(
             BUILD_CONFIG,
-            SCREEN_COMPOSABLES,
+            SCREEN_COMPOSABLES
         )
 
     val commonPackages: List<String>
@@ -48,7 +55,14 @@ sealed interface KoverExclusionRules {
             KOIN,
             DI,
             MODELS,
+            DOMAIN_MODELS,
             ROUTER_GRAPH,
+            EXTENSIONS,
+            CONFIGS,
+            TRANSFORMERS,
+            UTILS,
+            COMPONENTS,
+            PROVIDERS
         )
 
     val classes: List<String>
@@ -69,23 +83,10 @@ sealed interface KoverExclusionRules {
 
     object BusinessLogic : LogicModule {
         override val classes: List<String>
-            get() = commonClasses + listOf(
-                "eu.europa.ec.businesslogic.controller.security.AntiHookController",
-                "eu.europa.ec.businesslogic.controller.security.RootController",
-                "eu.europa.ec.businesslogic.controller.security.AndroidInstaller",
-                "eu.europa.ec.businesslogic.controller.security.AndroidPackageController*",
-                "eu.europa.ec.businesslogic.controller.security.AntiHookController*",
-                "eu.europa.ec.businesslogic.controller.security.RootControllerImpl",
-                "eu.europa.ec.businesslogic.controller.security.SecurityErrorCode",
-                "eu.europa.ec.businesslogic.controller.security.SecurityValidation",
-                "eu.europa.ec.businesslogic.extension.FlowExtensions*",
-                "eu.europa.ec.businesslogic.util.EudiWalletListenerWrapper",
-                "eu.europa.ec.businesslogic.util.SafeLet*",
-            )
+            get() = commonClasses
 
         override val packages: List<String>
             get() = commonPackages + listOf(
-                "eu.europa.ec.businesslogic.config",
                 "eu.europa.ec.businesslogic.controller.crypto",
                 "eu.europa.ec.businesslogic.controller.log",
                 "eu.europa.ec.businesslogic.controller.storage"
@@ -94,21 +95,14 @@ sealed interface KoverExclusionRules {
 
     object UiLogic : LogicModule {
         override val classes: List<String>
-            get() = commonClasses + listOf(
-                "eu.europa.ec.uilogic.navigation.*Screen*",
-                "eu.europa.ec.uilogic.navigation.ModuleRoute*",
-                "eu.europa.ec.uilogic.navigation.RouterHost*",
-                "eu.europa.ec.uilogic.serializer.UiSerializableParser*",
-                "eu.europa.ec.uilogic.serializer.UiSerializerImpl",
-            )
+            get() = commonClasses
 
         override val packages: List<String>
             get() = commonPackages + listOf(
-                "eu.europa.ec.uilogic.component",
-                "eu.europa.ec.uilogic.config",
                 "eu.europa.ec.uilogic.container",
-                "eu.europa.ec.uilogic.extension",
                 "eu.europa.ec.uilogic.mvi",
+                "eu.europa.ec.uilogic.navigation",
+                "eu.europa.ec.uilogic.serializer",
             )
     }
 
@@ -124,15 +118,10 @@ sealed interface KoverExclusionRules {
 
     object CommonFeature : FeatureModule {
         override val classes: List<String>
-            get() = commonClasses + listOf(
-                "eu.europa.ec.commonfeature.ui.document_details.DetailsContent*",
-            )
+            get() = commonClasses
 
         override val packages: List<String>
-            get() = commonPackages + listOf(
-                "eu.europa.ec.commonfeature.config",
-                "eu.europa.ec.commonfeature.ui.*.model",
-            )
+            get() = commonPackages
     }
 
     object StartupFeature : FeatureModule {
@@ -147,9 +136,7 @@ sealed interface KoverExclusionRules {
             get() = commonClasses
 
         override val packages: List<String>
-            get() = commonPackages + listOf(
-                "eu.europa.ec.dashboardfeature.ui.scanner.component",
-            )
+            get() = commonPackages
     }
 
     object PresentationFeature : FeatureModule {
@@ -165,9 +152,7 @@ sealed interface KoverExclusionRules {
             get() = commonClasses
 
         override val packages: List<String>
-            get() = commonPackages + listOf(
-                "eu.europa.ec.proximityfeature.ui.qr.component",
-            )
+            get() = commonPackages
     }
 
     object IssuanceFeature : FeatureModule {
